Payment cancel function implementation

+++ payment/models.py
@@ -19,7 +19,7 @@ class Payment(models.Model):
     )
     session_url = models.URLField(max_length=255)
     session_id = models.CharField(max_length=255)
-    money_to_pay = models.DecimalField(max_digits=5, decimal_places=2)
+    money_to_pay = models.IntegerField()
 
     def __str__(self) -> str:
         return f"{self.type} payment for borrowing {self.borrowing.id}"
